refactor: local storage for courses
