+++ src/hooks/index.ts
@@ -1,6 +1,5 @@
 import useDarkMode from './useDarkMode'
 import useMajor from './useMajor'
-import useCourses from './useCourses'
 import useShowGrid from './useShowGrid'
 
-export { useCourses, useMajor, useShowGrid, useDarkMode }
+export { useMajor, useShowGrid, useDarkMode }
